Fix build fails on Windows due enum cannot be treated as integer value error

+++ Source/Algorithm/AES/AES.cpp
@@ -43,7 +43,7 @@ String AES::Encrypt(const String& password, const String& buffer)
 	const String iv = Algorithm::GenerateRandomIV(CryptoPP::AES::BLOCKSIZE);
 
 	// Prepare key
-	CryptoPP::SecByteBlock key(CryptoPP::AES::MAX_KEYLENGTH + CryptoPP::AES::BLOCKSIZE); // Encryption key to be generated from user password + IV
+	CryptoPP::SecByteBlock key(static_cast<std::size_t>(CryptoPP::AES::MAX_KEYLENGTH) + static_cast<std::size_t>(CryptoPP::AES::BLOCKSIZE)); // Encryption key to be generated from user password + IV
 
 	// Convert key to KDF SHA-256, which allows you to use a password smaller or larger than the cipher's key size
 	CryptoPP::HKDF<CryptoPP::SHA256> hkdf{};
